change the behavior of Processor.apply() depending on the processor
