+++ src/rhoknp/processors/senter.py
@@ -1,6 +1,6 @@
 import logging
 import re
-from typing import List, Union
+from typing import List, Sequence, Union
 
 from rhoknp.processors.processor import Processor
 from rhoknp.units import Document, Sentence
@@ -23,6 +23,23 @@ class RegexSenter(Processor):
 
     def __repr__(self) -> str:
         return f"{self.__class__.__name__}()"
+
+    def apply(self, document: Union[Document, str]) -> Document:
+        """文書に解析器を適用する．
+
+        Args:
+            document (Union[Sentence, str]): 文書．
+        """
+        return self.apply_to_document(document)
+
+    def batch_apply(self, documents: Sequence[Union[Document, str]], processes: int = 0) -> List[Document]:
+        """複数文書に解析器を適用する．
+
+        Args:
+            documents (Sequence[Union[Document, str]]): 文書のリスト．
+            processes (int, optional): 並列処理数．0以下の場合はシングルプロセスで処理する．
+        """
+        return self.batch_apply_to_documents(documents, processes)
 
     def apply_to_document(self, document: Union[Document, str]) -> Document:
         """文書に RegexSenter を適用する．
